refactor: move mock data factories out of main file
